Rename controller.py to hub.py
