+++ homeassistant/components/unifi/button.py
@@ -30,7 +30,6 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .controller import UniFiController
 from .entity import (
     HandlerT,
     UnifiEntity,
@@ -38,6 +37,7 @@ from .entity import (
     async_device_available_fn,
     async_device_device_info_fn,
 )
+from .hub import UniFiController
 
 
 @callback
